# perf(git_children): optimize ref traversal