--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "graphite-cli",
-  "version": "0.10.1",
+  "version": "0.11.0",
   "license": "None",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",