**Details:**
Loading all refs into memory breaks if there are more than 6 million commits, and is very slow in general for large repos. The solution is to fetch subsets of refs for traversing git branch parents and children. In order to do this, Im dropping memoization for now because we're only loading subsets each time, but we could always bring back an advanced for later.




**Test Plan:**
Passes existing tests. I also profiled it against three repos:
cd graphite && gt ls = 4 sec before, 3.1 sec after
cd linux && gt ls =    25 sec before, 12 sec after
cd dagser && gt ls =   11 sec before, 5.5 sec after